requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
djoser~=2.2.3